src/main/java/com/anigenero/microservice/resource/json/serializer/LocalTimeSerializer.java
package com.anigenero.microservice.resource.json.serializer;

import com.fasterxml.jackson.core.JsonGenerator;
import com.fasterxml.jackson.databind.JsonSerializer;
import com.fasterxml.jackson.databind.SerializerProvider;

import java.io.IOException;
import java.time.LocalTime;

public class LocalTimeSerializer extends JsonSerializer<LocalTime> {

    @Override
    public void serialize(LocalTime value, JsonGenerator generator, SerializerProvider serializers)
            throws IOException {
        generator.writeNumber(value.toString());
    }

}
